+++ source/Public/Get-SqlDscDatabasePermission.ps1
@@ -15,6 +15,14 @@
         Specifies the name of the database principal for which the permissions are
         returned.
 
+    .PARAMETER Refresh
+        Specifies that the database's principal collections (Users, Roles, and
+        ApplicationRoles) should be refreshed before testing if the principal exists.
+        This is helpful when principals could have been modified outside of the
+        **ServerObject**, for example through T-SQL. But on databases with a large
+        amount of principals it might be better to make sure the **ServerObject**
+        is recent enough.
+
     .OUTPUTS
         [Microsoft.SqlServer.Management.Smo.DatabasePermissionInfo[]]
 
@@ -23,6 +31,13 @@
         Get-SqlDscDatabasePermission -ServerObject $serverInstance -DatabaseName 'MyDatabase' -Name 'MyPrincipal'
 
         Get the permissions for the principal 'MyPrincipal'.
+
+    .EXAMPLE
+        $serverInstance = Connect-SqlDscDatabaseEngine
+        Get-SqlDscDatabasePermission -ServerObject $serverInstance -DatabaseName 'MyDatabase' -Name 'MyPrincipal' -Refresh
+
+        Get the permissions for the principal 'MyPrincipal'. The database's principal
+        collections are refreshed before testing if the principal exists.
 
     .NOTES
         This command excludes fixed roles like _db_datareader_ by default, and will
@@ -52,7 +67,11 @@ function Get-SqlDscDatabasePermission
 
         [Parameter(Mandatory = $true)]
         [System.String]
-        $Name
+        $Name,
+
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $Refresh
     )
 
     # cSpell: ignore GSDDP
@@ -74,6 +93,11 @@ function Get-SqlDscDatabasePermission
                 DatabaseName      = $DatabaseName
                 Name              = $Name
                 ExcludeFixedRoles = $true
+            }
+
+            if ($Refresh.IsPresent)
+            {
+                $testSqlDscIsDatabasePrincipalParameters.Refresh = $true
             }
 
             $isDatabasePrincipal = Test-SqlDscIsDatabasePrincipal @testSqlDscIsDatabasePrincipalParameters